Location test additions
